requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
mysql-connector-python
bcrypt
ratelimit
openai
langfuse
captcha
mailersend
pyodbc
streamlit_google_auth
streamlit-extras
psycopg2-binary
pymupdf
streamlit-url-fragment